% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{change_door}
\alias{change_door}
\title{The contestant has a chance to change their door selection}
\usage{
change_door(stay = T, opened.door, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns the final selection.
}
\description{
\code{change_door(stay=T, opened.door, a.pick)} creates an if then statement that passes the information
from the previous selection to a function.
}
\details{
If the contestant stays its their final pick, if they want to switch
another door is selected.
}
\examples{
  change_door(stay=T, opened.door, a.pick)

}
