% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{determine_winner}
\alias{determine_winner}
\title{Determines if the contestant is a winner}
\usage{
determine_winner(final.pick, game)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns Win or Lose
}
\description{
\code{determine_winner(final.pick,game)} reads the final pick from the game.
from the previous selection to a function.
}
\details{
If the contestants door choice has a car paired with it then they win if not
they lose
}
\examples{
  determine_winner(final.pick,game)

}
