% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{open_goat_door}
\alias{open_goat_door}
\title{The host cannot select the car door and must start with a goat door.}
\usage{
open_goat_door(game, a.pick)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns selection from the contestant.
}
\description{
\code{open_goat_door(game, a.pick)} automatically assigns a goat door to the host
leaves the remaining door options to the contestant
}
\details{
Hosts selection can only be a goat door not a car door. the other 2 doors
are left for the contestant.
}
\examples{
  open_goat_door(game, a.pick)

}
