% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_game}
\alias{play_game}
\title{creates a strategy to categorize the amount of times switching results in a win or lose
and staying results in a win or lose.}
\usage{
play_game()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a table of strategy outcomes.
}
\description{
\code{play_game()}compiles the game data into a data frame
}
\details{
The data is randomized from the selections of doors and the option to stay
and switch from the stay or switch the game determines if the contestant wins or loses.
}
\examples{
  play_game()

}
