% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{play_n_games}
\alias{play_n_games}
\title{Determines how many times a simulated game can win or lose with a stay or switch strategy}
\usage{
play_n_games(n = 100)
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns a table of strategy outcomes and their probability.
}
\description{
\code{play_n_games(n = 100)}compiles the game data into a data frame and gives probability of chances
}
\details{
The data is randomized from the selections of doors and the option to stay
and switch from the stay or switch the game determines if the contestant wins or loses.
that data is compiled after a loop for 100 then that data is put into a table
the probability of the results are then displayed
}
\examples{
  play_n_games(n = 100)

}
