% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty-hall-pkg.R
\name{select_door}
\alias{select_door}
\title{select a door in the Monty Hall Problem game.}
\usage{
select_door()
}
\arguments{
\item{...}{no arguments are used by the function.}
}
\value{
The function returns one selection from the vector doors valued 1 through 3.
}
\description{
\code{select_door()} generates three doors a new game that consists of two doors
with goats behind them, and one with a car.
}
\details{
Before the contestant can open a door the doors must be generated
from a sample of three options in a vector called doors that variable.
}
\examples{
select_door()

}
